graph: leverage main covering-index and collation wins
After merging origin/main, pure Cypher count(*) over node scans lowers to
column-free SQL so core can use complete junction and registration indexes.
Document automatic vs graph-side leverage and pin ORDER BY / min-max text
semantics; adapt graph expand yield sites to the IOCompletions unit struct.

Tests: pure_count_star_over_*, pure_count_star_uses_junction_covering_index,
relationship_table_count_uses_registration_covering_index,
text_order_by_and_min_max_follow_sqlite_binary_collation

+++ core/vdbe/execute.rs
@@ -1405,7 +1405,7 @@ pub fn op_vfilter(
         cursor.filter(*idx_num as i32, idx_str, *arg_count, args)?
     };
     if step == crate::InternalVirtualTableStep::Yield {
-        return Ok(InsnFunctionStepResult::IO(IOCompletions::Single(
+        return Ok(InsnFunctionStepResult::IO(IOCompletions(
             Completion::new_yield(),
         )));
     }
@@ -1559,7 +1559,7 @@ pub fn op_vnext(
         cursor.next()?
     };
     if step == crate::InternalVirtualTableStep::Yield {
-        return Ok(InsnFunctionStepResult::IO(IOCompletions::Single(
+        return Ok(InsnFunctionStepResult::IO(IOCompletions(
             Completion::new_yield(),
         )));
     }
